page_art/requirements.txt
Pillow==10.0.0
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
openpyxl==3.1.5
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.1
psycopg2-binary==2.9.9